Move variables and add beta badge
