+++ apps/spruce/src/constants/requesters.ts
@@ -10,11 +10,15 @@ enum Requester {
   Trigger = "trigger_request",
 }
 
+const commitRequesters = [
+  Requester.AdHoc,
+  Requester.GitTag,
+  Requester.Gitter,
+  Requester.Trigger,
+];
+
 export const isMainlineRequester = (requester: Requester) =>
-  requester === Requester.AdHoc ||
-  requester === Requester.GitTag ||
-  requester === Requester.Gitter ||
-  requester === Requester.Trigger;
+  commitRequesters.includes(requester);
 
 const requesterToTitle: PartialRecord<Requester, string> = {
   [Requester.AdHoc]: "Periodic build",
@@ -31,4 +35,9 @@ const requesterToDescription: PartialRecord<Requester, string> = {
   [Requester.Trigger]: "Downstream trigger versions",
 };
 
-export { Requester, requesterToTitle, requesterToDescription };
+export {
+  Requester,
+  requesterToTitle,
+  requesterToDescription,
+  commitRequesters,
+};
